Moved heartland lib outside of pages. Added terms of use to paypal
