+++ src/components/checkout/TermsOfUseStatement.tsx
@@ -1,15 +1,25 @@
-'use client';
-export const TermsOfUseStatement = () => {
+import Link from 'next/link';
+
+type TermsOfUseStatementProps = {
+  isPaypal?: boolean;
+};
+export const TermsOfUseStatement = ({
+  isPaypal = false,
+}: TermsOfUseStatementProps) => {
+  const initialStatement = isPaypal
+    ? 'By checking out with Paypal'
+    : 'By clicking the "Submit Payment" button';
   return (
     <p className="pb-[40px] text-[14px] font-[400] text-[#767676] lg:border-t lg:border-[#DBDBDB] lg:pb-[48px] lg:pt-[35px]">
-      By clicking the “Submit Payment” button, you confirm that you have read,
-      understand, and accept our Terms of Use,{' '}
-      <a href="/policies/privacy-policy" className="underline">
-        Privacy Policy,
-      </a>{' '}
-      <a href="/policies/return-policy" className="underline">
+      {initialStatement}, you confirm that you have read, understand, and accept
+      our Terms of Use,{' '}
+      <Link href="/policies/privacy-policy" className="underline">
+        Privacy Policy
+      </Link>
+      ,{' '}
+      <Link href="/policies/return-policy" className="underline">
         Return Policy
-      </a>
+      </Link>
     </p>
   );
 };
